Remove TaroSky refrences
