+++ src/HM/FeedGenerator/Model/Singleton.php
@@ -5,7 +5,7 @@
  * @package HM\RSS_Delivery
  */
 
-namespace Tarosky\FeedGenerator\Model;
+namespace HM\FeedGenerator\Model;
 
 /**
  * Singleton Class
