+++ src/rss-delivery/service/HMExcite.php
@@ -8,8 +8,8 @@
 namespace HM\RSS_Delivery\Service;
 
 use HM\RSS_Delivery;
-use Tarosky\FeedGenerator\DeliveryManager;
-use Tarosky\FeedGenerator\Service\Excite;
+use HM\FeedGenerator\DeliveryManager;
+use HM\FeedGenerator\Service\Excite;
 use WP_Query;
 
 /**
