+++ src/rss-delivery/service/HMGoogleNews.php
@@ -7,8 +7,8 @@
 
 namespace HM\RSS_Delivery\Service;
 
-use Tarosky\FeedGenerator\DeliveryManager;
-use Tarosky\FeedGenerator\Service\GoogleNews;
+use HM\FeedGenerator\DeliveryManager;
+use HM\FeedGenerator\Service\GoogleNews;
 use WP_Query;
 
 /**
